refactor(react): move useValidation to src/hooks
also modifies the first parameter to allow a single ValidatorEntry
