+++ packages/react/src/hooks/index.ts
@@ -5,3 +5,4 @@ export * from './usePopper';
 export * from './usePopperTriggers';
 export * from './useSelect';
 export * from './useToken';
+export * from './useValidation';
